bin - server connection foundation
The foundation of the connection between the bin and the server are finalized, also, outline for the connection between the database and the server are placed. After the information about the db such as but not limited to the name of the table, what are the column names ect. are collected, the communication between the database and the server will be finalized.